make some fixes to improve test pass rate for lithium (#1886)

+++ testsuites/CBLTester/topology_specific_tests/sync_gateway_proxy_enabled/conftest.py
@@ -66,6 +66,7 @@ def params_from_base_suite_setup(request):
     delta_sync_enabled = request.config.getoption("--delta-sync")
     cbs_ce = request.config.getoption("--cbs-ce")
     sg_ce = request.config.getoption("--sg-ce")
+    cbs_ssl = request.config.getoption("--server-ssl")
     cluster_config = request.config.getoption("--cluster-config")
     enable_encryption = request.config.getoption("--enable-encryption")
     encryption_password = request.config.getoption("--encryption-password")
@@ -163,6 +164,15 @@ def params_from_base_suite_setup(request):
     else:
         log_info("Running without delta sync")
         persist_cluster_config_environment_prop(cluster_config, 'delta_sync_enabled', False)
+
+    if cbs_ssl:
+        log_info("Running tests with cbs <-> sg ssl enabled")
+        # Enable ssl in cluster configs
+        persist_cluster_config_environment_prop(cluster_config, 'cbs_ssl_enabled', True)
+    else:
+        log_info("Running tests with cbs <-> sg ssl disabled")
+        # Disable ssl in cluster configs
+        persist_cluster_config_environment_prop(cluster_config, 'cbs_ssl_enabled', False)
 
     if hide_product_version:
         log_info("Suppress the SGW product Version")
@@ -370,7 +380,8 @@ def params_from_base_suite_setup(request):
         "encryption_password": encryption_password,
         "cbs_ce": cbs_ce,
         "sg_ce": sg_ce,
-        "prometheus_enable": prometheus_enable
+        "prometheus_enable": prometheus_enable,
+        "ssl_enabled": cbs_ssl
     }
 
     if request.node.testsfailed != 0 and enable_file_logging and create_db_per_suite is not None:
@@ -444,6 +455,7 @@ def params_from_base_test_setup(request, params_from_base_suite_setup):
     cbs_ce = params_from_base_suite_setup["cbs_ce"]
     sg_ce = params_from_base_suite_setup["sg_ce"]
     prometheus_enable = request.config.getoption("--prometheus-enable")
+    cbs_ssl = params_from_base_suite_setup["ssl_enabled"]
 
     source_db = None
     test_name_cp = test_name.replace("/", "-")
@@ -538,6 +550,7 @@ def params_from_base_test_setup(request, params_from_base_suite_setup):
         "test_cbllog": test_cbllog,
         "cbs_ce": cbs_ce,
         "sg_ce": sg_ce,
+        "ssl_enabled": cbs_ssl,
         "prometheus_enable": prometheus_enable
     }
 
@@ -652,6 +665,23 @@ def setup_customized_teardown_test(params_from_base_test_setup):
         "cbl_db3": cbl_db3,
     }
     log_info("Tearing down test")
-    db.deleteDB(cbl_db1)
-    db.deleteDB(cbl_db2)
-    db.deleteDB(cbl_db3)
+    path = db.getPath(cbl_db1).rstrip("/\\")
+    path2 = db.getPath(cbl_db2).rstrip("/\\")
+    path3 = db.getPath(cbl_db3).rstrip("/\\")
+    if '\\' in path:
+        path = '\\'.join(path.split('\\')[:-1])
+        path2 = '\\'.join(path2.split('\\')[:-1])
+        path3 = '\\'.join(path3.split('\\')[:-1])
+    else:
+        path = '/'.join(path.split('/')[:-1])
+        path2 = '/'.join(path2.split('/')[:-1])
+        path3 = '/'.join(path3.split('/')[:-1])
+    try:
+        if db.exists(cbl_db_name1, path):
+            db.deleteDB(cbl_db1)
+        if db.exists(cbl_db_name2, path2):
+            db.deleteDB(cbl_db2)
+        if db.exists(cbl_db_name3, path3):
+            db.deleteDB(cbl_db3)
+    except Exception as err:
+        log_info("Exception occurred: {}".format(err))
